Sketch code that would be used when all customers have migrated
